requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
fastapi==0.89.1
Flask==2.2.2
frozenlist==1.3.3
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
mypy-extensions==0.4.3
openai==0.26.1
pathspec==0.10.3
platformdirs==2.6.2
pydantic==1.10.4
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.2
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
tqdm==4.64.1
twilio==7.16.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
Werkzeug==2.2.2
yarl==1.8.2
